\name{Kmedians-package}
\alias{Kmedians-package}
%\alias{Gmedian}
\docType{package}
\title{
\packageTitle{Kmedians}
}
\description{
We provide functions to provide robust clustering. Function \code{\link{gen_K}} enables to generate a sample of data following a contaminated Gaussian mixture. Functions \code{\link{Kmedians}} and \code{\link{Kmeans}} consists in a K-median and a K-means algorithms while \code{\link{Kplot}} enables to produce graph for both methods.
}

\author{
\packageAuthor{Kmedians}

Maintainer: \packageMaintainer{Kmedians}
}
\references{
Cardot, H., Cenac, P. and Zitt, P-A. (2013). Efficient and fast estimation of the geometric median in Hilbert spaces with an averaged stochastic gradient algorithm. \emph{Bernoulli}, 19, 18-43.

Cardot, H. and Godichon-Baggioni, A. (2017). Fast Estimation of the Median Covariation Matrix with Application to Online Robust Principal Components Analysis.  \emph{Test}, 26(3), 461-480

Vardi, Y. and Zhang, C.-H. (2000). The multivariate L1-median and associated data depth. \emph{Proc. Natl. Acad. Sci. USA}, 97(4):1423-1426.

}
\keyword{Robust clustering}
